display graphql nodes; extracted menu in a separate component; versions badge
